[add]Storeの追加

+++ app/src/main/java/jp/okmt/notepad/store/MemoStore.kt
@@ -6,7 +6,7 @@ import jp.okmt.notepad.memo.MemoIndex
 interface MemoStore {
     val memoList: List<MemoIndex>
 
-    fun read(id: Int): Memo
+    fun read(id: Long): Memo
 
     fun write(memo: Memo)
